Redesign diagnosis list
