+++ src/Components/Facility/ConsultationDetails/index.tsx
@@ -39,8 +39,7 @@ import PatientNotesSlideover from "../PatientNotesSlideover";
 import { AssetBedModel } from "../../Assets/AssetTypes";
 import PatientInfoCard from "../../Patient/PatientInfoCard";
 import RelativeDateUserMention from "../../Common/RelativeDateUserMention";
-import LegacyDiagnosesList from "../../Diagnosis/LegacyDiagnosesList";
-import AccordionV2 from "../../Common/components/AccordionV2";
+import DiagnosesListAccordion from "../../Diagnosis/DiagnosesListAccordion";
 
 const Loading = lazy(() => import("../../Common/Loading"));
 const PageTitle = lazy(() => import("../../Common/PageTitle"));
@@ -404,20 +403,9 @@ export const ConsultationDetails = (props: any) => {
         </div>
         <div className="col-span-1 mt-2 overflow-hidden rounded-lg bg-white shadow">
           <div className="px-4 py-2">
-            <AccordionV2
-              expanded={true}
-              title={
-                <h3 className="text-lg font-semibold leading-relaxed text-gray-900">
-                  Diagnoses
-                </h3>
-              }
-            >
-              <div className="mt-2">
-                <LegacyDiagnosesList
-                  diagnoses={consultationData.diagnoses || []}
-                />
-              </div>
-            </AccordionV2>
+            <DiagnosesListAccordion
+              diagnoses={consultationData.diagnoses ?? []}
+            />
           </div>
         </div>
         <div className="mt-4 w-full border-b-2 border-gray-200">
